chore: move DAGs into dedicated `CodeEntropy/levels/graph` module
