+++ CodeEntropy/levels/graph/level_dag.py
@@ -12,11 +12,11 @@ from typing import Any
 import networkx as nx
 
 from CodeEntropy.levels.axes import AxesCalculator
-from CodeEntropy.levels.conformation_dag import ConformationDAG
 from CodeEntropy.levels.execution.policy import ExecutionPolicy
 from CodeEntropy.levels.execution.reducers import NeighborReducer
 from CodeEntropy.levels.execution.scheduler import FrameScheduler
-from CodeEntropy.levels.frame_dag import FrameGraph
+from CodeEntropy.levels.graph.conformation_dag import ConformationDAG
+from CodeEntropy.levels.graph.frame_dag import FrameGraph
 from CodeEntropy.levels.neighbors import Neighbors
 from CodeEntropy.levels.nodes.accumulators import InitCovarianceAccumulatorsNode
 from CodeEntropy.levels.nodes.axes_topology import BuildAxesTopologyNode
